Some comments in the code

+++ video_fft/src/main.cpp
@@ -51,8 +51,7 @@ inline torch::Tensor do_pad_video(torch::Tensor & tensor_input) {
 }
 
 extern "C" {
-int do_fft_compress(void *blob, int len_t, int len_y, int len_x, int len_c,
-                    float fps, float freq_limit, void *dest) {
+int do_fft_compress(void *blob, uint16_t len_t, uint16_t len_y, uint16_t len_x, uint8_t len_c, float fps, float freq_limit, void *dest) {
 
   int64_t dist_c = 1;
   int64_t dist_x = len_c * dist_c;
